+++ tools4a-fragment-ktx/src/androidTest/java/com/github/panpf/tools4a/fragment/ktx/test/FragmentxTest.kt
@@ -96,7 +96,7 @@ class FragmentxTest {
 
     @Test
     fun testInstantiate() {
-        val fragment = Fragment::class.java.instantiate(BundleBuilder().putString("key", "testInstantiate").build())
+        val fragment = Fragment::class.java.instantiate(bundleOf("key" to "testInstantiate"))
         assertEquals(Fragment::class.java.name, fragment.javaClass.name)
         assertEquals("testInstantiate", fragment.arguments!!.getString("key"))