+++ tools4a-fragment-ktx/src/main/java/androidx/fragment/app/ktx/FragmentInsider.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright (C) 2020 panpf <panpfpanpf@outlook.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+@file:Suppress("NOTHING_TO_INLINE")
+
+package androidx.fragment.app.ktx
+
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentInsider
+import androidx.lifecycle.Lifecycle
+
+/**
+ * Returns the maxState of the specified Fragment
+ */
+inline fun Fragment.getMaxState(): Lifecycle.State = FragmentInsider.getMaxState(this)
+
+/**
+ * Returns the smallest maxState of the specified Fragment and all its parent Fragments
+ */
+inline fun Fragment.getMaxStateWithParent(): Lifecycle.State = FragmentInsider.getMaxStateWithParent(this)