+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * change: Refactor the names of all methods in Packagex
 * change: Fragmentx added isUserVisibleHintWithParent, isResumedWithParent, findUserVisibleFragmentRecursivelyByUserVisibleHint, findUserVisibleFragmentRecursivelyByResumed method
 * change: Fragmentx's findUserVisibleFragmentByUserVisibleHint, findUserVisibleFragmentByResumed methods no longer traverse grandchildren Fragment
+* new: Added FragmentInsider class, which contains getMaxState, getMaxStateWithParent methods
 
 ## v1.0.0-rc01